refactor: member domain, DTO, Repository 패키지 변경
