+++ src/main/java/com/example/demo/domain/StudyMember.java
@@ -14,6 +14,8 @@ import javax.persistence.ManyToOne;
 import javax.persistence.MapsId;
 import javax.persistence.Table;
 
+import com.example.demo.domain.member.MemberImpl;
+
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
