+++ src/main/java/com/example/demo/domain/StudyOnceImpl.java
@@ -21,6 +21,8 @@ import javax.persistence.OneToMany;
 import javax.persistence.Table;
 
 import com.example.demo.domain.cafe.CafeImpl;
+import com.example.demo.domain.member.Member;
+import com.example.demo.domain.member.MemberImpl;
 import com.example.demo.exception.CafegoryException;
 import com.example.demo.exception.ExceptionType;
