+++ src/main/java/com/example/demo/domain/member/Member.java
@@ -1,6 +1,8 @@
-package com.example.demo.domain;
+package com.example.demo.domain.member;
 
 import java.util.List;
+
+import com.example.demo.domain.StudyMember;
 
 public interface Member {
 	Long getId();
