+++ src/main/java/com/example/demo/domain/member/ThumbnailImage.java
@@ -1,4 +1,4 @@
-package com.example.demo.domain;
+package com.example.demo.domain.member;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
