+++ src/main/java/com/example/demo/dto/member/MemberResponse.java
@@ -1,4 +1,4 @@
-package com.example.demo.dto;
+package com.example.demo.dto.member;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
