+++ src/main/java/com/example/demo/mapper/StudyMemberMapper.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import com.example.demo.domain.StudyMember;
-import com.example.demo.dto.MemberResponse;
+import com.example.demo.dto.member.MemberResponse;
 
 public class StudyMemberMapper {
