+++ src/main/java/com/example/demo/repository/StudyMemberRepositoryCustom.java
@@ -3,8 +3,8 @@ package com.example.demo.repository;
 import java.time.LocalDate;
 import java.util.List;
 
-import com.example.demo.domain.MemberImpl;
 import com.example.demo.domain.StudyMember;
+import com.example.demo.domain.member.MemberImpl;
 
 public interface StudyMemberRepositoryCustom {
 	List<StudyMember> findByMemberAndStudyDate(MemberImpl member, LocalDate studyDate);
