+++ src/main/java/com/example/demo/repository/member/MemberRepository.java
@@ -1,10 +1,10 @@
-package com.example.demo.repository;
+package com.example.demo.repository.member;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.demo.domain.MemberImpl;
+import com.example.demo.domain.member.MemberImpl;
 
 public interface MemberRepository extends JpaRepository<MemberImpl, Long> {
 	Optional<MemberImpl> findByEmail(String email);
