+++ src/test/java/com/example/demo/builder/TestStudyMemberBuilder.java
@@ -1,9 +1,9 @@
 package com.example.demo.builder;
 
 import com.example.demo.domain.Attendance;
-import com.example.demo.domain.MemberImpl;
 import com.example.demo.domain.StudyMember;
 import com.example.demo.domain.StudyOnceImpl;
+import com.example.demo.domain.member.MemberImpl;
 
 public class TestStudyMemberBuilder {
