+++ src/test/java/com/example/demo/builder/TestStudyOnceCommentBuilder.java
@@ -1,8 +1,8 @@
 package com.example.demo.builder;
 
-import com.example.demo.domain.MemberImpl;
 import com.example.demo.domain.StudyOnceComment;
 import com.example.demo.domain.StudyOnceImpl;
+import com.example.demo.domain.member.MemberImpl;
 
 public class TestStudyOnceCommentBuilder {
 	private Long id;
