+++ src/test/java/com/example/demo/helper/StudyOnceCommentPersistHelper.java
@@ -4,9 +4,9 @@ import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
 import com.example.demo.builder.TestStudyOnceCommentBuilder;
-import com.example.demo.domain.MemberImpl;
 import com.example.demo.domain.StudyOnceComment;
 import com.example.demo.domain.StudyOnceImpl;
+import com.example.demo.domain.member.MemberImpl;
 
 public class StudyOnceCommentPersistHelper {
