+++ src/test/java/com/example/demo/repository/ReviewRepositoryTest.java
@@ -16,10 +16,10 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.example.demo.config.QueryDslConfig;
 import com.example.demo.config.TestConfig;
-import com.example.demo.domain.MemberImpl;
 import com.example.demo.domain.ReviewImpl;
-import com.example.demo.domain.ThumbnailImage;
 import com.example.demo.domain.cafe.CafeImpl;
+import com.example.demo.domain.member.MemberImpl;
+import com.example.demo.domain.member.ThumbnailImage;
 import com.example.demo.helper.CafePersistHelper;
 import com.example.demo.helper.MemberPersistHelper;
 import com.example.demo.helper.ReviewPersistHelper;
